refactor PlantCareDescription
